requirements.txt
python-telegram-bot
typer[all]
python-dotenv
rich
